(android) Make fee/liquidity warning more visible in Receive
The previous warning message was easy to miss, yet not concise
enough, without bringing enough information.

When the app detects that an invoice may not be received, a
full screen dialog may be displayed. Whether this dialog is
displayed or not depends on the information available. We may
just display a one-line warning, which shows more information
when clicked -- though this one-line warning is more visible
with a /!\ icon.

A message is also displayed when a fee is expected and within
the bounds of the user configuration. This message use a
different wording and icon.

The wording has been reworked to provide actionable solutions,
with a button saying "Increase the fee limit", or a hint to
the "request liquidity" feature.
