+++ phoenix-android/src/main/res/values-es/important_strings.xml
@@ -113,10 +113,19 @@
     <string name="receive_lightning_warning_title_mayfail">El pago puede fracasar</string>
     <string name="receive_lightning_warning_title_fee_expected">Tasa prevista</string>
 
-    <string name="receive_lightning_warning_fee_insufficient_liquidity">La liquidez entrante es insuficiente.</string>
-    <string name="receive_lightning_warning_fee_policy_disabled_insufficient_liquidity">La liquidez entrante es insuficiente y ha desactivado la gestión automatizada de canales.\nPuntee para configurarlo.</string>
-    <string name="receive_lightning_warning_fee_within_policy">Puede ser necesaria una tasa de %1$s (≈ %2$s) para recibir este pago.\nPuntee para configurarlo.</string>
-    <string name="receive_lightning_warning_fee_exceeds_policy">La tasa será probablemente superior a su límite máximo (%1$s).\nPuntee para configurarlo.</string>
+    <string name="receive_lightning_warning_hint_show_details">Más información</string>
+
+    <string name="receive_lightning_sheet_dismiss">Cerrar</string>
+    <string name="receive_lightning_sheet_button_enable">Activar la gestión automátizada</string>
+    <string name="receive_lightning_sheet_button_configure">Configurar tasas máximas</string>
+
+    <string name="receive_lightning_warning_message_fee_expected">Probablemente será necesaria una transacción en la cadena para recibir esta cantidad.\n\nLas comisiones serán de aproximadamente %1$s.</string>
+    <string name="receive_lightning_warning_message_fee_expected_unknown">Probablemente será necesaria una transacción en la cadena para recibir esta cantidad.</string>
+
+    <string name="receive_lightning_warning_message_above_limit">Se estima una tasa de %1$s por recibir este importe, y esta tasa excede su límite autorizado de %2$s.\n\nAumente este límite o solicite liquidez adicional.</string>
+    <string name="receive_lightning_warning_message_above_limit_percent">Se estiman unas tasas de %1$s para recibir este importe, y estas tasas superan el %2$s%% del importe.\n\nAumente este límite o solicite liquidez adicional..</string>
+
+    <string name="receive_lightning_warning_fee_policy_disabled_insufficient_liquidity">La liquidez entrante es insuficiente y tienes desactivada la gestión automatizada de canales.</string>
 
     <string name="receive_onchain_legacy_warning_title">Aviso de migración</string>
     <string name="receive_onchain_legacy_warning">Esta es su nueva dirección. No reutilice sus antiguas direcciones Bitcoin anteriores a la migración.</string>
